fix: create student view
